backend/requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv